python_backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
argcomplete==1.10.3
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.28.0
atlassian-python-api==3.41.3
attrs
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
compressed-rtf==1.0.6
dataclasses-json==0.5.14
DateTime==5.2
Deprecated==1.2.14
distro==1.8.0
docx2txt==0.8
ebcdic==1.1.1
exceptiongroup==1.1.3
extract-msg==0.28.7
fastapi==0.103.2
frozenlist==1.4.0
fsspec==2023.9.2
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
greenlet==3.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.6
IMAPClient==2.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.314
langsmith==0.0.43
litellm==1.7.11
llama-hub==0.0.43
llama-index==0.9.10
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.1
oauthlib==3.2.2
olefile==0.46
openai==1.1.0
packaging==23.2
pandas==2.1.1
pdfminer==20191125
pdfminer.six==20191110
pdftotext==2.2.2
pgvector==0.2.3
Pillow==9.5.0
priority==2.0.0
prisma==0.10.0
protobuf==4.24.4
psutil==5.9.6
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.22
pythonnet==3.0.3
pytz==2023.3.post1
PyYAML==6.0.1
ratelimiter==1.2.0.post0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
six==1.12.0
slack-sdk==3.23.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.21
starlette==0.27.0
taskgroup==0.0.0a4
tenacity==8.2.3
git+git://github.com/3chospirits/textract.git
tiktoken==0.5.1
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.17
uuid==1.30
uvicorn==0.23.2
wrapt==1.15.0
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.1.9
yarl==1.9.3
zope.interface==6.1
